refactor(examples): isolate root and leaf renders; Examples as FC
- main.tsx is now just bootstrap (glob + lazy root/leaf renders); no mvc
  Component or adapter import at the top level
- Move organize() and route types into _layout/loader.ts
- Examples becomes a plain FC deriving routes from the modules glob, with a
  bare @expressive/react import to register the adapter for the shell chunk
